# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sankara Eye Donation Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /units:
    get:
      operationId: listUnits
      tags: [units]
      summary: List all Sankara hospital units
      responses:
        "200":
          description: List of units
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Unit"
    post:
      operationId: createUnit
      tags: [units]
      summary: Create a new hospital unit (Admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnitInput"
      responses:
        "201":
          description: Created unit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unit"
        "400":
          description: Invalid input
        "401":
          description: Unauthorized
  /units/{id}:
    get:
      operationId: getUnit
      tags: [units]
      summary: Get a specific unit by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Unit details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unit"
        "404":
          description: Unit not found
    patch:
      operationId: updateUnit
      tags: [units]
      summary: Update a unit (Admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnitUpdate"
      responses:
        "200":
          description: Updated unit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unit"
        "404":
          description: Unit not found
    delete:
      operationId: deleteUnit
      tags: [units]
      summary: Delete a unit (Admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Unit not found

  # ─── USERS ──────────────────────────────────────────────────────────────────
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (Admin only)
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AppUser"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user (Admin only — all users must be created by MHQ admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppUser"
        "400":
          description: Invalid input
        "401":
          description: Unauthorized
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a specific user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppUser"
        "404":
          description: User not found
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user (Admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppUser"
        "404":
          description: User not found
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user (Admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: User not found
  /users/{id}/toggle-active:
    patch:
      operationId: toggleUserActive
      tags: [users]
      summary: Activate or deactivate a user (Admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppUser"
        "404":
          description: User not found
  /users/{id}/reset-password:
    post:
      operationId: resetUserPassword
      tags: [users]
      summary: Reset a user's password (Admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordReset"
      responses:
        "200":
          description: Password reset successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: User not found

  # ─── EYE CALLS ──────────────────────────────────────────────────────────────
...
  /public/eye-calls:
    post:
      operationId: submitPublicEyeCall
      tags: [public]
      summary: Submit public eye donation form (no auth required)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublicEyeCallInput"
      responses:
        "201":
          description: Submitted successfully, includes WhatsApp message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicEyeCallResponse"
        "400":
          description: Invalid input
...
components:
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # ─── UNIT ─────────────────────────────────────────────────────────────────
    Unit:
      type: object
      required: [id, name, state, district, coordinatorName, coordinatorWhatsapp, coordinatorEmail, isActive, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        state:
          type: string
        district:
          type: string
        address:
          type: ["string", "null"]
        coordinatorName:
          type: string
        coordinatorWhatsapp:
          type: string
        coordinatorEmail:
          type: string
          format: email
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UnitInput:
      type: object
      required: [name, state, district, coordinatorName, coordinatorWhatsapp, coordinatorEmail]
      properties:
        name:
          type: string
          minLength: 1
        state:
          type: string
          minLength: 1
        district:
          type: string
          minLength: 1
        address:
          type: string
        coordinatorName:
          type: string
          minLength: 1
        coordinatorWhatsapp:
          type: string
          minLength: 10
        coordinatorEmail:
          type: string
          format: email
          minLength: 1
        isActive:
          type: boolean
          default: true
    UnitUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        state:
          type: string
        district:
          type: string
        address:
          type: string
        coordinatorName:
          type: string
        coordinatorWhatsapp:
          type: string
        coordinatorEmail:
          type: string
          format: email
        isActive:
          type: boolean
...
    AppUser:
      type: object
      required: [id, clerkId, email, name, role, isActive, mustChangePassword, createdAt, updatedAt]
      properties:
        id:
          type: integer
        clerkId:
          type: string
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [super_admin, eye_bank_head, unit_coordinator]
        unitId:
          type: ["integer", "null"]
        unitName:
          type: ["string", "null"]
        isActive:
          type: boolean
        mustChangePassword:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    UserInput:
      type: object
      required: [email, name, role]
      properties:
        email:
          type: string
          format: email
        name:
          type: string
          minLength: 1
        role:
          type: string
          enum: [super_admin, eye_bank_head, unit_coordinator]
        unitId:
          type: ["integer", "null"]
        password:
          type: string
          minLength: 8
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
          enum: [super_admin, eye_bank_head, unit_coordinator]
        unitId:
          type: ["integer", "null"]
        isActive:
          type: boolean
...
    PasswordReset:
      type: object
      required: [newPassword]
      properties:
        newPassword:
          type: string
          minLength: 8
...
    EyeCall:
      type: object
      required: [id, callId, donorName, donorAge, donorGender, timeOfDeath, causeOfDeath, referrerName, referrerMobile, referrerRelationship, state, district, pincode, address, unitId, unitName, status, createdAt, updatedAt]
      properties:
        id:
          type: integer
        callId:
          type: string
        donorName:
          type: string
        donorAge:
          type: integer
        donorGender:
          type: string
          enum: [male, female, other]
        timeOfDeath:
          type: string
        causeOfDeath:
          type: string
        referrerName:
          type: string
        referrerMobile:
          type: string
        referrerRelationship:
          type: string
        state:
          type: string
        district:
          type: string
        pincode:
          type: string
        address:
          type: string
        unitId:
          type: integer
        unitName:
          type: string
        status:
          type: string
          enum: [new, contacted, team_sent, completed, cancelled]
        notes:
          type: ["string", "null"]
        thirdPartyHospitalDetails:
          type: ["string", "null"]
        thirdPartyHelperContact:
          type: ["string", "null"]
        attendedByUserId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    PublicEyeCallInput:
      type: object
      required: [donorName, donorAge, donorGender, timeOfDeath, causeOfDeath, referrerName, referrerMobile, referrerRelationship, state, district, pincode, address, unitId]
      properties:
        donorName:
          type: string
          minLength: 1
        donorAge:
          type: integer
          minimum: 0
          maximum: 120
        donorGender:
          type: string
          enum: [male, female, other]
        timeOfDeath:
          type: string
        causeOfDeath:
          type: string
        referrerName:
          type: string
          minLength: 1
        referrerMobile:
          type: string
          minLength: 10
        referrerRelationship:
          type: string
        state:
          type: string
        district:
          type: string
        pincode:
          type: string
        address:
          type: string
        unitId:
          type: integer
    PublicEyeCallResponse:
      type: object
      required: [eyeCall, whatsappMessage, whatsappUrl]
      properties:
        eyeCall:
          $ref: "#/components/schemas/EyeCall"
        whatsappMessage:
          type: string
        whatsappUrl:
          type: string

    # ─── DASHBOARD ────────────────────────────────────────────────────────────
...